Script that will walk a SQL Server database to enable un-trusted & disabled foreign keys
This script will enable keys that are un-trusted and disabled, in the
event of an error due to data inconsistency it will capture the rows
into an XML format and the T-SQL statement to find them.

This information is stored in a local session temp table, with the
results being selected out to the caller (SSMS/SQLCMD) so that they can
be seen.